+++ src/features/manageProfileSettings/components/ProfileSettings/ProfileSettings.tsx
@@ -3,6 +3,7 @@ import { FormattedMessage } from 'react-intl';
 
 import classNames from 'classnames';
 
+import { PROFILE_SETTINGS_DEFAULT_OPTION } from '@/common/constants/profileSettings.constants';
 import { StatusType } from '@/common/constants/status.constants';
 import { UserNotificationFactory } from '@/common/services/userNotification';
 import { Button, ButtonType } from '@/components/Button';
@@ -14,9 +15,9 @@ import { useLoadingState, useManageProfileSettingsState, useStatusState, useUISt
 
 import ArrowLeftIcon from '@/assets/arrow-left-16.svg?react';
 
-import { CustomProfileToggle } from '../CustomProfileToggle';
 import { DefaultProfileOption } from '../DefaultProfileOption';
 import { ProfileSettingsEditor } from '../ProfileSettingsEditor';
+import { ProfileSettingsList } from '../ProfileSettingsList';
 
 import './ProfileSettings.scss';
 
@@ -24,11 +25,14 @@ export const ProfileSettings = () => {
   const { setIsLoading } = useLoadingState();
   const { loadProfile } = useLoadProfile();
   const { loadProfileSettings } = useLoadProfileSettings();
-  const { selectedProfile, setFullProfile, setProfileSettings } = useManageProfileSettingsState([
-    'selectedProfile',
-    'setFullProfile',
-    'setProfileSettings',
-  ]);
+  const { selectedProfile, selectedProfileSettingsMeta, setFullProfile, setProfileSettings, resetProfileSettings } =
+    useManageProfileSettingsState([
+      'selectedProfile',
+      'selectedProfileSettingsMeta',
+      'setFullProfile',
+      'setProfileSettings',
+      'resetProfileSettings',
+    ]);
   const {
     isManageProfileSettingsBelowBreakpoint,
     isManageProfileSettingsShowEditor,
@@ -54,9 +58,20 @@ export const ProfileSettings = () => {
           setIsLoading(true);
           const profile = await loadProfile(selectedProfile.id);
           setFullProfile(profile);
-          setProfileSettings(
-            await loadProfileSettings(String(selectedProfile.id), profile, selectedProfile.resourceType),
-          );
+          if (selectedProfileSettingsMeta) {
+            if (selectedProfileSettingsMeta.id === PROFILE_SETTINGS_DEFAULT_OPTION) {
+              resetProfileSettings();
+            } else {
+              setProfileSettings(
+                await loadProfileSettings(
+                  selectedProfileSettingsMeta.id,
+                  String(selectedProfile.id),
+                  profile,
+                  selectedProfile.resourceType,
+                ),
+              );
+            }
+          }
         } catch {
           addStatusMessagesItem?.(
             UserNotificationFactory.createMessage(StatusType.error, 'ld.errorLoadingProfileSettings'),
@@ -68,7 +83,7 @@ export const ProfileSettings = () => {
 
       initialize();
     }
-  }, [selectedProfile]);
+  }, [selectedProfile, selectedProfileSettingsMeta]);
 
   const showView =
     !isManageProfileSettingsBelowBreakpoint ||
@@ -96,7 +111,7 @@ export const ProfileSettings = () => {
 
       <hr />
 
-      <CustomProfileToggle />
+      <ProfileSettingsList />
 
       <ProfileSettingsEditor />
     </div>